default qt quick 2.0 project init
These files are the result of starting a new Qt Quick 2.0 application in
QtCreator. The path goes kind of deep, but when you build on your
system, it's going to put artifacts at the same level as 'src'. This
keeps the code away from any other resources we want to store in the
root.